Video texture test  * use libvpx for video decoding  * async update textures

+++ Time.cpp
@@ -6,25 +6,25 @@ namespace Sys {
 std::chrono::steady_clock::time_point init_time = std::chrono::steady_clock::now();
 }
 
-uint32_t Sys::GetTimeMs() {
+uint64_t Sys::GetTimeMs() {
     auto t = (std::chrono::steady_clock::now() - init_time);
     auto tt = std::chrono::duration_cast<std::chrono::milliseconds>(t);
-    return (uint32_t)tt.count();
+    return uint64_t(tt.count());
 }
 
 uint64_t Sys::GetTimeUs() {
     auto t = (std::chrono::steady_clock::now() - init_time);
     auto tt = std::chrono::duration_cast<std::chrono::microseconds>(t);
-    return (uint64_t)tt.count();
+    return uint64_t(tt.count());
 }
 
 uint64_t Sys::GetTimeNs() {
     auto t = (std::chrono::steady_clock::now() - init_time);
     auto tt = std::chrono::duration_cast<std::chrono::nanoseconds>(t);
-    return (uint64_t)tt.count();
+    return uint64_t(tt.count());
 }
 
 double Sys::GetTimeS() {
     std::chrono::duration<double> t = (std::chrono::steady_clock::now() - init_time);
-    return (double)t.count();
+    return double(t.count());
 }
