+++ Time_.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 namespace Sys {
-uint32_t GetTimeMs();
+uint64_t GetTimeMs();
 uint64_t GetTimeUs();
 uint64_t GetTimeNs();
 double GetTimeS();
